pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "spectacle-toolbelt"
version = "0.1.0"
description = "A companion extension kit for KDE Spectacle."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "GPL-3.0-or-later" }
authors = [
  { name = "Ryan Winkler" }
]
keywords = ["kde", "spectacle", "screenshot", "scrolling-capture", "imagemagick"]
classifiers = [
  "Development Status :: 2 - Pre-Alpha",
  "Environment :: X11 Applications :: KDE",
  "Intended Audience :: End Users/Desktop",
  "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
  "Programming Language :: Python :: 3",
  "Topic :: Multimedia :: Graphics :: Capture :: Screen Capture",
  "Topic :: Utilities"
]
dependencies = [
  "Pillow>=10.0",
  "websockets>=12.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0"
]

[project.scripts]
spectacle-toolbelt = "spectacle_toolbelt.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
